update scripts and ush/scripts

+++ scripts/plots/mesoscale/exevs_mesoscale_sref_grid2obs_plots.sh
@@ -93,7 +93,8 @@ for fcst_valid_hour in $fcst_valid_hours ; do
    VARs='CAPEsfc DPT2m TCDC'
    score_types='lead_average threshold_average'
   else
-   err_exit "$stats is wrong stat"
+   echo $stats is wrong stat
+   exit
   fi   
 
   for score_type in $score_types ; do
@@ -266,7 +267,7 @@ for var in hgt tmp ugrd vgrd prmsl rh dpt tcdc cape ; do
 	unit=m
 	var_level=${var}_${level}m
     elif [ $level = L0 ] ; then
-	unit=''
+	unit='non'
 	var_level=${var}_l0
     fi
 
@@ -287,11 +288,14 @@ for var in hgt tmp ugrd vgrd prmsl rh dpt tcdc cape ; do
 	 fi
 
       
-         if [ $unit = mb ] || [ $unit = m ] ; then	    
-            mv ${score_type}_regional_conus_valid_${valid}_${level}${unit}_${var}_${stat}${end}  evs.sref.${stat}.${var_level}.last${past_days}days.${scoretype}_valid_${valid}.buk_conus.png
+         if [ $unit = mb ] || [ $unit = m ] ; then	
+	   if [ -s ${score_type}_regional_conus_valid_${valid}_${level}${unit}_${var}_${stat}${end} ] ; then	 
+             mv ${score_type}_regional_conus_valid_${valid}_${level}${unit}_${var}_${stat}${end}  evs.sref.${stat}.${var_level}.last${past_days}days.${scoretype}_valid_${valid}.buk_conus.png
+	   fi
          else
-           mv ${score_type}_regional_conus_valid_${valid}_${var}_${stat}${end}  evs.sref.${stat}.${var_level}.last${past_days}days.${scoretype}_valid_${valid}.buk_conus.png
- 
+           if [ -s ${score_type}_regional_conus_valid_${valid}_${var}_${stat}${end} ] ; then 
+             mv ${score_type}_regional_conus_valid_${valid}_${var}_${stat}${end}  evs.sref.${stat}.${var_level}.last${past_days}days.${scoretype}_valid_${valid}.buk_conus.png
+	   fi
          fi
                
       else
@@ -301,11 +305,14 @@ for var in hgt tmp ugrd vgrd prmsl rh dpt tcdc cape ; do
 	   end='f6_to_f87.png'
 
            if [ $unit = m ] ; then
-	    	   
-	     mv ${score_type}_regional_conus_valid_${valid}_${level}${unit}_${var}_${stat}_${end}  evs.sref.${stat}.${var_level}.last${past_days}days.${scoretype}_valid_${valid}.buk_conus.png
-	   else
-             mv ${score_type}_regional_conus_valid_${valid}_${var}_${stat}_${end}  evs.sref.${stat}.${var_level}.last${past_days}days.${scoretype}_valid_${valid}.buk_conus.png
-           fi
+	      if [ -s ${score_type}_regional_conus_valid_${valid}_${level}${unit}_${var}_${stat}_${end} ] ; then  	   
+	        mv ${score_type}_regional_conus_valid_${valid}_${level}${unit}_${var}_${stat}_${end}  evs.sref.${stat}.${var_level}.last${past_days}days.${scoretype}_valid_${valid}.buk_conus.png
+              fi
+           else
+	      if [ -s ${score_type}_regional_conus_valid_${valid}_${var}_${stat}_${end} ] ; then
+                mv ${score_type}_regional_conus_valid_${valid}_${var}_${stat}_${end}  evs.sref.${stat}.${var_level}.last${past_days}days.${scoretype}_valid_${valid}.buk_conus.png
+              fi
+   	   fi
          fi
       fi
